Set up path and version managers in ~/.zshrc, not ~/.zshenv
This ensures the version managers (rbenv, nodenv, dinghy, etc.) continue to work on OS X 10.11 “El Capitan.” 

The issue was that in OS X 10.11, the invocation of `path_helper` was moved from /etc/zshenv to /etc/zprofile in the system ZSH config, which meant that our local ~/.zshenv ran before the main system paths were set, which meant that it couldn’t find the installations of rbenv et al., so they were never initialised.

Thoughtbot is encouraging a different change, to actually move those system configs around (https://github.com/thoughtbot/dotfiles/commit/15de0bcfee0d7a6472da048029408c5faedbae4e), but I’d rather leave them where they are, especially when its debatable as to whether ~/.zshenv is the right place to do this kind of setup (see https://github.com/thoughtbot/dotfiles/pull/426 for an ongoing discussion about this).

The fix at this point is just to move all of this setup to the top of ~/.zshrc.